+++ jib-core/src/main/java/com/google/cloud/tools/jib/cache/CacheStorageFiles.java
@@ -153,7 +153,7 @@ class CacheStorageFiles {
    */
   Path getImageDirectory(ImageReference imageReference) {
     // Replace ':' and '@' with '!' to avoid directory-naming restrictions
-    String replacedReference = imageReference.toStringWithTag().replace(':', '!').replace('@', '!');
+    String replacedReference = imageReference.toStringWithIdentifier().replace(':', '!').replace('@', '!');
 
     // Split image reference on '/' to build directory structure
     Iterable<String> directories = Splitter.on('/').split(replacedReference);